Dev improve testflow

+++ QPCore/Controllers/AppFeatureController.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using QPCore.Model.DataBaseModel;
 using Microsoft.AspNetCore.Cors;
 using QPCore.Service;
+using QPCore.Model.Common;
+using Microsoft.AspNetCore.Authorization;
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace QPCore.Controllers
@@ -10,7 +12,8 @@ namespace QPCore.Controllers
     [Route("api/[controller]")]
     //[EnableCors("AnyOrignPolicy")]
     [ApiController]
-    public class AppFeatureController : ControllerBase
+    [Authorize]
+    public class AppFeatureController : BaseController
     {
         private FeatureAppService _featureService;
         /// <summary>
@@ -39,26 +42,52 @@ namespace QPCore.Controllers
 
         // POST api/<AppFeatureController>
         [HttpPost]
-        public AppFeatureView Post(AppFeatureView feature)
+        public ActionResult<AppFeatureView> Post (AppFeatureView feature)
         {
+            var isExitedName = _featureService.CheckFeatureNameExisted(feature.FeatureName, feature.ParentFeatureId);
+            if (isExitedName)
+            {
+                return BadRequest(new BadRequestResponse()
+                {
+                    Message = string.Format(CommonMessageList.EXISTED_NAME_STRING, feature.FeatureName)
+                });
+            }
             AppFeatureView appfeature = _featureService.CreateFeature(feature);
-            return appfeature;
+            return Ok(appfeature);
         }
 
         // PUT api/<AppFeatureController>/5
         [HttpPut("{id}")]
-        public AppFeatureView Put(int id, AppFeatureView feature)
+        public ActionResult<AppFeatureView> Put(int id, AppFeatureView feature)
         {
             feature.AppFeatureId = id;
+
+            var isExitedName = _featureService.CheckFeatureNameExisted(feature.FeatureName, feature.ParentFeatureId, feature.AppFeatureId);
+            if (isExitedName)
+            {
+                return BadRequest(new BadRequestResponse()
+                {
+                    Message = string.Format(CommonMessageList.EXISTED_NAME_STRING, feature.FeatureName)
+                });
+            }
             AppFeatureView appfeature = _featureService.UpdateFeature(feature);
-            return appfeature;
+            return Ok(appfeature);
         }
 
         // DELETE api/<AppFeatureController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public ActionResult Delete(int id)
         {
+            var canDelete = _featureService.CheckCanDelete(id);
+            if (!canDelete)
+            {
+                return BadRequest(new BadRequestResponse()
+                {
+                    Message = "Can not delete item which is included step which step source is code"
+                });
+            }
             _featureService.DeleteFeature(id);
+            return Ok();
         }
     }
 }
